Move and extract general class for VowlObjects
