+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/HasAnnotations.java
@@ -1,0 +1,15 @@
+/*
+ * HasIri.java
+ *
+ */
+
+package de.uni_stuttgart.vis.vowl.owl2vowl.model;
+
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.annotation.Annotations;
+
+/**
+ *
+ */
+public interface HasAnnotations {
+	Annotations getAnnotations();
+}
