+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/HasIri.java
@@ -1,0 +1,15 @@
+/*
+ * HasIri.java
+ *
+ */
+
+package de.uni_stuttgart.vis.vowl.owl2vowl.model;
+
+import org.semanticweb.owlapi.model.IRI;
+
+/**
+ *
+ */
+public interface HasIri {
+	IRI getIri();
+}
