+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/data/VowlData.java
@@ -5,13 +5,13 @@
 
 package de.uni_stuttgart.vis.vowl.owl2vowl.model.data;
 
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.AbstractEntity;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.AbstractNode;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes.AbstractClass;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.datatypes.AbstractDatatype;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.properties.AbstractProperty;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.properties.VowlDatatypeProperty;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.properties.VowlObjectProperty;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.AbstractEntity;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.AbstractNode;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes.AbstractClass;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.datatypes.AbstractDatatype;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.properties.AbstractProperty;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.properties.VowlDatatypeProperty;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.properties.VowlObjectProperty;
 import org.semanticweb.owlapi.model.IRI;
 
 import java.util.*;
