+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/HasEquivalents.java
@@ -1,4 +1,4 @@
-package de.uni_stuttgart.vis.vowl.owl2vowl.model;
+package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities;
 
 import org.semanticweb.owlapi.model.IRI;
