+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/nodes/HasComplement.java
@@ -1,4 +1,4 @@
-package de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes;
+package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes;
 
 import org.semanticweb.owlapi.model.IRI;
