+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/nodes/classes/HasDisjointUnion.java
@@ -3,7 +3,7 @@
  *
  */
 
-package de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes;
+package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes;
 
 import org.semanticweb.owlapi.model.IRI;
