+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/nodes/classes/HasIndividuals.java
@@ -1,4 +1,4 @@
-package de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes;
+package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes;
 
 import org.semanticweb.owlapi.model.IRI;
