+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/model/entities/properties/HasInverse.java
@@ -1,4 +1,4 @@
-package de.uni_stuttgart.vis.vowl.owl2vowl.model.properties;
+package de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.properties;
 
 import org.semanticweb.owlapi.model.IRI;
