+++ src/main/java/de/uni_stuttgart/vis/vowl/owl2vowl/parser/vowl/property/DomainRangeVisitor.java
@@ -1,10 +1,9 @@
 package de.uni_stuttgart.vis.vowl.owl2vowl.parser.vowl.property;
 
 import de.uni_stuttgart.vis.vowl.owl2vowl.model.data.VowlData;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.data.VowlGenerationEnum;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes.AbstractClass;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes.NullClass;
-import de.uni_stuttgart.vis.vowl.owl2vowl.model.nodes.classes.VowlClass;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes.AbstractClass;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes.NullClass;
+import de.uni_stuttgart.vis.vowl.owl2vowl.model.entities.nodes.classes.VowlClass;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.semanticweb.owlapi.model.*;
